backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic==1.10.8
alembic==1.11.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0 